+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## [Unreleased]
+### Changed
+- `StepReporter.sendStep` methods now returns `Maybe<String>` with Step ID, by @HardNorth
+- `StepReporter.finishNestedStep` methods now returns `Maybe<OperationCompletionRS>`, by @HardNorth
+- `rxjava` version updated on 2.2.21, by @HardNorth
 
 ## [5.2.25]
 ### Changed